+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+  - Deployment - Mixed jobs shouldn't set deployment to
+    error state on first try
   - docker-compose-cli connector substitue env variables
     to avoid failures during start listing containers
   - Execute_command separate stdout from stderr